Consolidate types into FHIRDefinitions
