+++ src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './FHIRDefinitions';
-export * from './Fishable';
 export * from './logger';
